+++ packages/core/src/components/Table/Body/Row/Row.styled.tsx
@@ -1,0 +1,43 @@
+import { styled } from '@medly-components/utils';
+import { rgba } from 'polished';
+import { GridTemplateProps } from '../../types';
+import { Props } from './types';
+
+export const Row = styled('tr').attrs(({ gridTemplateColumns }: GridTemplateProps) => ({
+    style: {
+        gridTemplateColumns
+    }
+}))<Props>`
+    min-width: fit-content;
+    display: grid;
+    align-items: center;
+    cursor: ${({ disabled, onClick }) => (disabled ? 'not-allowed' : onClick ? 'pointer' : 'inherit')};
+
+    &:hover {
+        z-index: 2;
+        box-shadow: ${({ disabled, onClick, theme }) => !disabled && onClick && `0 0.2rem 0.4rem ${rgba(theme.table.shadowColor, 0.2)} `};
+    }
+
+    &:nth-child(odd) {
+        & > * {
+            background-color: ${({ theme, isSelected, disabled }) =>
+                theme.table.row.bgColor[disabled ? 'disabled' : isSelected ? 'selected' : 'odd']};
+        }
+    }
+
+    &:nth-child(even) {
+        & > * {
+            background-color: ${({ theme, isSelected, disabled }) =>
+                theme.table.row.bgColor[disabled ? 'disabled' : isSelected ? 'selected' : 'even']};
+        }
+    }
+
+    &:not(:last-child) {
+        border-bottom: 1px solid ${({ theme }) => theme.table.borderColor};
+    }
+`;
+
+export const NoResult = styled(Row)`
+    text-align: center;
+    line-height: 30px;
+`;