+++ packages/core/src/components/Table/Body/types.ts
@@ -1,14 +1,15 @@
-import { ColumnConfig, Data } from '../types';
+import { Data, TableColumnConfig } from '../types';
 
 export interface Props {
-    data: Data[];
-    columns: ColumnConfig[];
-    selectedRows?: number[];
-    uniqueKeyName: string;
+    data: Data;
+    columns: TableColumnConfig[];
+    selectedRowIds?: number[];
+    rowIdentifier: string;
     rowSelectionDisableKey?: string;
     rowClickDisableKey?: string;
     isLoading?: boolean;
     addColumnMaxSize: (field: string, value: number) => void;
     onRowSelection?: (id: number) => void;
     onRowClick?: (rowData: object) => void;
+    showShadowAfterFrozenElement?: boolean;
 }