+++ packages/core/src/components/Table/Row/types.ts
@@ -1,4 +1,0 @@
-import { HTMLProps } from '@medly-components/utils';
-import { GridTemplateProps } from '../types';
-
-export type Props = GridTemplateProps & HTMLProps<HTMLLIElement>;